Added content panes
These panes enable the user to go back and forward through the panes.

ContentPane is the main abstract class that one must extend if he plans
to add an additional pane.

Added the MainPain which, as the name suggests, is the main pane of the
application.

The main window class has been renamed into MainWindow, for more
clarity, and less text